chore: re-organized some api/redux code
