+++ src/pages/types/index.jsx
@@ -4,7 +4,7 @@ import Loader from "~/components/loader";
 import Error from "~/components/error";
 import TitleBar from "~/components/title-bar";
 import TypeCard from "./type-card";
-import { useServicesIndex } from "~/core/store/services/useServices";
+import { useServicesIndex } from "~/core/api/services/useServices";
 
 const TypesList = styled.ul({
   listStyle: "none",
